remove unused scripts in gaia example
